[Refactor] artifact-type-definition.cpp/h をsystem/ からsystem/artifact/ へ移動させた
