+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -1538,9 +1538,6 @@
     <ClCompile Include="..\..\src\artifact\fixed-art-generator.cpp">
       <Filter>artifact</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\system\artifact-type-definition.cpp">
-      <Filter>system</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\view\display-store.cpp">
       <Filter>view</Filter>
     </ClCompile>
@@ -2570,6 +2567,9 @@
     </ClCompile>
     <ClCompile Include="..\..\src\system\floor\town-records.cpp">
       <Filter>system\floor</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\system\artifact\artifact-definition.cpp">
+      <Filter>system\artifact</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -4334,9 +4334,6 @@
     <ClInclude Include="..\..\src\artifact\fixed-art-generator.h">
       <Filter>artifact</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\system\artifact-type-definition.h">
-      <Filter>system</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\view\display-store.h">
       <Filter>view</Filter>
     </ClInclude>
@@ -5583,6 +5580,9 @@
     <ClInclude Include="..\..\src\system\floor\town-records.h">
       <Filter>system\floor</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\system\artifact\artifact-definition.h">
+      <Filter>system\artifact</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\src\wall.bmp" />
@@ -5882,6 +5882,9 @@
     </Filter>
     <Filter Include="external-lib\fmt">
       <UniqueIdentifier>{de5173c3-9151-46e9-a937-8861a5a14c00}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="system\artifact">
+      <UniqueIdentifier>{58714fb7-9c67-4cd4-b69f-a5b0d7d7964f}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
