+++ src/combat/shoot.cpp
@@ -44,7 +44,7 @@
 #include "player/player-skill.h"
 #include "player/player-status-table.h"
 #include "sv-definition/sv-bow-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-key.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/enums/terrain/terrain-characteristics.h"
