+++ src/flavor/flavor-util.cpp
@@ -4,7 +4,7 @@
 #include "object-enchant/tr-types.h"
 #include "object/tval-types.h"
 #include "sv-definition/sv-food-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/item-entity.h"
 #include "util/string-processor.h"
