+++ src/floor/floor-object.cpp
@@ -23,7 +23,7 @@
 #include "object/object-kind-hook.h"
 #include "perception/object-perception.h"
 #include "range/v3/range/conversion.hpp"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-allocation.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
