+++ src/info-reader/artifact-reader.cpp
@@ -8,7 +8,7 @@
 #include "locale/japanese.h"
 #include "main/angband-headers.h"
 #include "object-enchant/tr-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "util/bit-flags-calculator.h"
 #include "util/enum-converter.h"
 #include "util/string-processor.h"
