+++ src/info-reader/general-parser.cpp
@@ -10,7 +10,7 @@
 #include "player-info/class-types.h"
 #include "player-info/race-types.h"
 #include "realm/realm-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/building-type-definition.h"
