+++ src/knowledge/knowledge-items.cpp
@@ -18,7 +18,7 @@
 #include "object/tval-types.h"
 #include "perception/identification.h"
 #include "perception/object-perception.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/floor/floor-info.h"
