+++ src/monster-floor/special-death-switcher.cpp
@@ -27,7 +27,7 @@
 #include "sv-definition/sv-other-types.h"
 #include "sv-definition/sv-protector-types.h"
 #include "sv-definition/sv-weapon-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-allocation.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
