+++ src/object-enchant/item-magic-applier.cpp
@@ -12,7 +12,7 @@
 #include "object-enchant/object-curse.h"
 #include "object-enchant/special-object-flags.h"
 #include "player/player-status-flags.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/floor/floor-info.h"
