+++ src/racial/racial-android.cpp
@@ -14,7 +14,7 @@
 #include "sv-definition/sv-armor-types.h"
 #include "sv-definition/sv-protector-types.h"
 #include "sv-definition/sv-weapon-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "target/target-getter.h"
