+++ src/save/save.cpp
@@ -27,7 +27,7 @@
 #include "save/lore-writer.h"
 #include "save/player-writer.h"
 #include "save/save-util.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-info.h"
 #include "system/floor/town-list.h"
