+++ src/specific-object/bloody-moon.cpp
@@ -4,7 +4,7 @@
 #include "object-enchant/tr-types.h"
 #include "player-base/player-race.h"
 #include "racial/racial-android.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
