+++ src/spell/spells-object.cpp
@@ -31,7 +31,7 @@
 #include "sv-definition/sv-other-types.h"
 #include "sv-definition/sv-scroll-types.h"
 #include "sv-definition/sv-weapon-types.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/baseitem/baseitem-key.h"
 #include "system/floor/floor-info.h"
 #include "system/item-entity.h"
