+++ src/wizard/artifact-analyzer.cpp
@@ -6,7 +6,7 @@
 #include "object-enchant/trc-types.h"
 #include "object-enchant/trg-types.h"
 #include "object/object-info.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/item-entity.h"
 #include "term/z-form.h"
 #include "util/bit-flags-calculator.h"
