+++ src/wizard/wizard-special-process.cpp
@@ -54,7 +54,7 @@
 #include "spell-kind/spells-world.h"
 #include "spell/spells-status.h"
 #include "status/bad-status-setter.h"
-#include "system/artifact-type-definition.h"
+#include "system/artifact/artifact-definition.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
